moved generate function
